sentinel_ai/requirements.txt
google-genai
pyautogui
Pillow
python-dotenv
pygetwindow
speechrecognition
pyttsx3
pyaudio
pytesseract
plyer
